Moved json back to env (couldn't change branches)
