--- config/saml.config.json
@@ -1,6 +1,0 @@
-{
-  "attributeMap": {},
-  "decryptionPvkPath": "",
-  "privateKey": "",
-  "cert": []
-}
